split map
