+++ luxonis_train/attached_modules/metrics/__init__.py
@@ -2,7 +2,6 @@ from .base_metric import BaseMetric
 from .confusion_matrix import ConfusionMatrix
 from .embedding_metrics import ClosestIsPositiveAccuracy, MedianDistances
 from .mean_average_precision import MeanAveragePrecision
-from .mean_average_precision_keypoints import MeanAveragePrecisionKeypoints
 from .object_keypoint_similarity import ObjectKeypointSimilarity
 from .torchmetrics import Accuracy, F1Score, JaccardIndex, Precision, Recall
 
@@ -12,7 +11,6 @@ __all__ = [
     "JaccardIndex",
     "BaseMetric",
     "MeanAveragePrecision",
-    "MeanAveragePrecisionKeypoints",
     "ObjectKeypointSimilarity",
     "Precision",
     "Recall",
